WebHost: allow getting checksum-specific datapackage via /api/datapackage/<checksum>
